4.21

+++ osa4/bloglist/controllers/blogs.js
@@ -61,8 +61,26 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async(request, response) =>{
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  console.log("token: ", decodedToken)
+
+  if (!decodedToken.id) {
+    return response.status(401).json({error: 'token invalid'})
+  }
+  const blog = await Blog.findById(request.params.id)
+  const user = await User.findById(decodedToken.id)
+
+  if (!(blog && user)){
+    response.status(404).end()
+  }
+  
+  if ( blog.user.toString() === user._id.toString() ) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(403).end()
+  }
 })
